support heredoc in mirb
as #1564 described, if we type "<<EOF" in mirb, it would fail in a
segfault instead of enter a opened code block, and unfortunately #1565
did not really resolve this issue.

the root cause of this is heredocs does not been parsed immediately,
but delayed in the next line. this is marked by p->heredoces_from_nextline,
but is_code_block_open() in mirb only checks p->parsing_heredoc.

the segfault from `mruby -e "<<"` is another issue, should be addressed
in a different way.